Kart Selection Scene and General UI
Created a new scene that is a menu in which players select their kart. An "Animations" and "Sprites" folder was created for the UI elements, and TextMeshPro was also installed. A basic script was done that updates the kart model's material when the player selects a different option (though this will be replaced by a different prefab instead of a different material).